backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.7.4
pydantic-settings==2.3.4
google-generativeai==0.8.3
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==23.2.1
websockets==12.0
python-multipart==0.0.9
